Rule clean up before edit support (#478)
Restructure the parsing module and tidy up some loose ends.
