+++ crates/rules/src/object.rs
@@ -7,10 +7,8 @@
  */
 
 use std::fmt::{Display, Formatter};
-use std::str::FromStr;
 
-use crate::parse::StrTrace;
-use crate::{bool_to_c, parse, ObjPart, Rvalue};
+use crate::{bool_to_c, ObjPart, Rvalue};
 
 /// # Object
 /// The object is the file that the subject is interacting with.
@@ -104,17 +102,6 @@ impl Display for Part {
             Part::FileType(t) => f.write_fmt(format_args!("ftype={}", t)),
             Part::Path(p) => f.write_fmt(format_args!("path={}", p)),
             Part::Trust(b) => f.write_fmt(format_args!("trust={}", bool_to_c(*b))),
-        }
-    }
-}
-
-impl FromStr for Object {
-    type Err = String;
-
-    fn from_str(i: &str) -> Result<Self, Self::Err> {
-        match parse::object(StrTrace::new(i)) {
-            Ok((_, s)) => Ok(s),
-            Err(_) => Err("Failed to parse Object from string".into()),
         }
     }
 }
