+++ crates/rules/src/parser/mod.rs
@@ -8,8 +8,18 @@
 
 pub mod errat;
 pub mod error;
+pub mod from_str;
 pub mod legacy;
+pub mod parse;
 pub mod trace;
+
+pub mod comment;
+pub mod decision;
+pub mod object;
+pub mod permission;
+pub mod rule;
+pub mod set;
+pub mod subject;
 
 #[cfg(test)]
 mod tests;
