+++ crates/rules/src/permission.rs
@@ -7,10 +7,6 @@
  */
 
 use std::fmt::{Display, Formatter};
-use std::str::FromStr;
-
-use crate::parse;
-use crate::parse::StrTrace;
 
 /// # Permission
 /// Describes what kind permission is being asked for. The permission is either
@@ -35,17 +31,6 @@ impl Display for Permission {
             Permission::Any => f.write_str("any"),
             Permission::Open => f.write_str("open"),
             Permission::Execute => f.write_str("execute"),
-        }
-    }
-}
-
-impl FromStr for Permission {
-    type Err = String;
-
-    fn from_str(i: &str) -> Result<Self, Self::Err> {
-        match parse::permission(StrTrace::new(i)) {
-            Ok((_, s)) => Ok(s),
-            Err(_) => Err("Failed to parse Permission from string".into()),
         }
     }
 }
